requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
APScheduler==3.10.4
python-telegram-bot==22.5
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
